refactor(): adjust the project layout
Adjust the project layout.
Changed the term "系列" to "专题".
