+++ README.md
@@ -34,18 +34,18 @@ KouShare-dl 是一个使用 [Cobra](https://github.com/spf13/cobra)
   * [三、下载视频](#三下载视频)
     + [3.1 使用默认参数下载视频](#31-使用默认参数下载视频)
     + [3.2 下载视频至指定文件夹](#32-下载视频至指定文件夹)
-    + [3.3 下载某个系列的所有视频](#33-下载某个系列的所有视频)
+    + [3.3 下载某个专题的所有视频](#33-下载某个专题的所有视频)
     + [3.4 下载不同清晰度的视频](#34-下载不同清晰度的视频)
   * [四、录制直播与下载快速回放](#四录制直播与下载快速回放)
     + [4.1 对指定直播间进行录制](#41-对指定直播间进行录制)
     + [4.2 合并录制的视频片段](#42-合并录制的视频片段)
     + [4.3 下载直播间快速回放视频](#43-下载直播间快速回放视频)
   * [五、下载课件](#五下载课件)
-    + [5.1 下载单个课件和系列课件](#51-下载单个课件和系列课件)
+    + [5.1 下载单个课件和专题课件](#51-下载单个课件和专题课件)
     + [5.2 优化 pdf 文件【实验性功能】](#52-优化-pdf-文件实验性功能)
 - [FAQ](#faq)
     - [KouShare-dl 下载视频时是并行下载吗？](#koushare-dl-下载视频时是并行下载吗)
-    - [下载系列视频时因网络波动导致下载中断该怎么办？](#下载系列视频时因网络波动导致下载中断该怎么办)
+    - [下载专题视频时因网络波动导致下载中断该怎么办？](#下载专题视频时因网络波动导致下载中断该怎么办)
     - [下载视频的过程中遇到因被占用而导致文件重命名失败的错误应该如何处理？](#下载视频的过程中遇到因被占用而导致文件重命名失败的错误应该如何处理)
 - [鸣谢](#鸣谢)
 
@@ -57,7 +57,7 @@ KouShare-dl 是一个使用 [Cobra](https://github.com/spf13/cobra)
 
 - 获取视频或直播的详细信息
 
-- 下载单个蔻享视频或整个系列的视频
+- 下载单个蔻享视频或整个专题的视频
 
 - 下载清晰度为标清、高清和超清的视频（需要登录）
 
@@ -71,7 +71,7 @@ KouShare-dl 是一个使用 [Cobra](https://github.com/spf13/cobra)
 
 - 下载直播间的快速回放🚀
 
-- 下载单个课件或整个系列的课件
+- 下载单个课件或整个专题的课件
 
 ### 它无法做到的事情
 
@@ -79,7 +79,7 @@ KouShare-dl 是一个使用 [Cobra](https://github.com/spf13/cobra)
 
 ### "功能支持"表格
 
-| 类型 | 是否支持系列下载 | 是否支持单独下载 | 是否支持断点续传 | 是否支持不同清晰度的下载 | 是否支持付费产品下载 |
+| 类型 | 是否支持专题下载 | 是否支持单独下载 | 是否支持断点续传 | 是否支持不同清晰度的下载 | 是否支持付费产品下载 |
 | :--: | :--------------: | :--------------: | :--------------: | :----------------------: | :------------------: |
 | 视频 |        ✔️         |        ✔️         |        ✔️         |            ✔️             |          ⭕           |
 | 直播 |        ➖         |        ✔️         |        ✔️         |            ❌             |          ➖           |
@@ -103,7 +103,7 @@ cd C:\Users\lenovo\Downloads\
 ks version
 ```
 
-若出现`KouShare-dl v0.8.4`字样，则说明可以正常使用。接下来您可以继续输入 KouShare-dl 程序的命令来进行交互。比如，输入`ks help`并回车，您就可以看到 KouShare-dl 程序的帮助信息了。
+若出现`KouShare-dl v0.8.5`字样，则说明可以正常使用。接下来您可以继续输入 KouShare-dl 程序的命令来进行交互。比如，输入`ks help`并回车，您就可以看到 KouShare-dl 程序的帮助信息了。
 
 # 命令简介
 
@@ -140,7 +140,7 @@ KouShare-dl 程序的命令具有下面的格式：
   -q, --quality    指定下载视频的清晰度（high为超清，standard为高清，low为标清，不指定则默认为超清）
       --qpdf-bin   指定qpdf的bin文件夹所在的路径（注：该flag无简写形式）
   -r, --replay     指定是否下载直播间快速回放视频
-  -s, --series     指定是否下载整个系列的文件
+  -s, --series     指定是否下载整个专题的文件
   -v, --version    查看版本号
 ```
 
@@ -178,7 +178,7 @@ ks logout
 
 **查看视频信息**使用`ks info [vid]`命令。`info`命令没有 flag 。
 
-执行该命令后会返回指定 vid 的视频的详细信息，包括标题、讲者、单位、日期、时长、体积、类别、系列、分组以及视频简介等。
+执行该命令后会返回指定 vid 的视频的详细信息，包括标题、讲者、单位、日期、时长、体积、类别、专题、分组以及视频简介等。
 
 几点说明：
 
@@ -186,7 +186,7 @@ ks logout
 
 - 若“体积”为`0MB [未知]`，则说明该视频是未购买的（或未在购买有效期内的）付费视频，此时 KouShare-dl 无法获取该视频的体积信息。
 
-- 若“系列”不为空，说明该视频是属于某个系列的视频，比如某精品公开课中的一节课。
+- 若“专题”不为空，说明该视频是属于某个专题的视频，比如某精品公开课中的一节课。
 
 
 您可以试一试下面的例子：
@@ -221,7 +221,7 @@ https://www.koushare.com/video/videodetail/7412
 | :------: | :---------: | :------------------: | :------: | :----------: |
 |   `-p`   |  `--path`   |  指定保存视频的路径  | `String` | 当前所在路径 |
 |   `-q`   | `--quality` | 指定下载视频的清晰度 | `String` |     超清     |
-|   `-s`   | `--series`  | 指定是否下载系列视频 |  `Bool`  |      否      |
+|   `-s`   | `--series`  | 指定是否下载专题视频 |  `Bool`  |      否      |
 
 多个 flag 可以不分顺序地叠加使用，但`Bool`类型的 flag 宜放在最后使用。关于命令中 flag 的详细使用语法，可以参考[这里的描述](https://github.com/spf13/pflag#command-line-flag-syntax)。
 
@@ -255,17 +255,17 @@ ks save 7552 --path=D:\temp\
 
 若指定的文件夹不存在，程序会创建该文件夹以保存视频。若遇到`Access is denied`的错误提示，则说明权限不足，此时您需要使用更高的权限来运行 KouShare-dl。
 
-### 3.3 下载某个系列的所有视频
+### 3.3 下载某个专题的所有视频
 
-系列下载需要指定`-s`参数，`-s`或`--series`参数是`Bool`型 flag，使用时无需指定具体的值。
+专题下载需要指定`-s`参数，`-s`或`--series`参数是`Bool`型 flag，使用时无需指定具体的值。
 
-您需要知道所要下载的系列视频中任意一个视频的 vid。以“中物院研究生院精品公开课之《高等量子力学》公开课程”系列为例，可以使用下面这条命令下载该系列的所有视频：
+您需要知道所要下载的专题视频中任意一个视频的 vid。以“中物院研究生院精品公开课之《高等量子力学》公开课程”专题为例，可以使用下面这条命令下载该专题的所有视频：
 
 ```shell
 ks save 7304 -s
 ```
 
-程序会使用该系列的名字创建一个文件夹用以存放下载的视频。`7304`是该系列第一个视频的 vid，可被替换为该系列任意视频的 vid。
+程序会使用该专题的名字创建一个文件夹用以存放下载的视频。`7304`是该专题第一个视频的 vid，可被替换为该专题任意视频的 vid。
 
 若要同时指定保存视频的位置（如`D:\temp\`），可以运行该命令：
 
@@ -407,9 +407,9 @@ ks live 447482 -r --path="C:\Users\lenovo\Desktop"
 | :------: | :----------: | :----------------------------: | :------: | :----------: |
 |   `-p`   |   `--path`   |       指定保存课件的路径       | `String` | 当前所在路径 |
 |    无    | `--qpdf-bin` | 指定qpdf的bin文件夹所在的路径  | `String` |  不使用qpdf  |
-|   `-s`   |  `--series`  | 指定是否下载整个系列的所有课件 |  `Bool`  |      否      |
+|   `-s`   |  `--series`  | 指定是否下载整个专题的所有课件 |  `Bool`  |      否      |
 
-### 5.1 下载单个课件和系列课件
+### 5.1 下载单个课件和专题课件
 
 下载课件时不需要处于登录状态下。假如您想要下载为 vid 为`7405`的视频关联的课件，可以运行该命令：
 
@@ -417,15 +417,15 @@ ks live 447482 -r --path="C:\Users\lenovo\Desktop"
 ks slide 7405
 ```
 
-使用`info`命令查看 vid 为`7405`的视频信息，可以发现该视频的“系列”不为空，说明该视频还有其它相关视频。
+使用`info`命令查看 vid 为`7405`的视频信息，可以发现该视频的“专题”不为空，说明该视频还有其它相关视频。
 
-假如您想下载这个系列视频的所有课件，可以使用`-s`参数：
+假如您想下载这个专题视频的所有课件，可以使用`-s`参数：
 
 ```shell
 ks slide 7405 -s
 ```
 
-同样地，`7405`可以被替换为同系列任意视频的 vid。
+同样地，`7405`可以被替换为同专题任意视频的 vid。
 
 ### 5.2 优化 pdf 文件【实验性功能】
 
@@ -442,7 +442,7 @@ ks slide 7405 --qpdf-bin=C:\Downloads\qpdf-10.1.0\bin\
 #### KouShare-dl 下载视频时是并行下载吗？
 不是并行下载。
 
-#### 下载系列视频时因网络波动导致下载中断该怎么办？
+#### 下载专题视频时因网络波动导致下载中断该怎么办？
 再次运行您上一次使用的下载命令，KouShare-dl 会自动跳过已下载完成的视频，并继续完成您的下载。
 录制直播意外中断时同理。
